Move the Loader class to its own file
According to the PSR-4 and PSR-12 standards, a class should
be in a file of its own. This also allows for more
flexibility.

The src directory doesn't follow the PSR-4 standard yet.
